Email LDU and SPO
When the SPO manually creates a case allocation, we will send an email to the LDU and a copy
is also sent to the SPO. We only send these emails if the offender's probation service will
be in Wales or England. Cases where the probation service is in Scotland or Northern Ireland
are handled outside of this application.

When an offender's case information is updated, we will send an email to the LDU and SPO if
the LDU is different from the LDU recorded previously.
Upon create action (only for wales and english cases)
Upon update action but only when cases update team which has a different ldu than previously